Modify: Restructure Frontend Component and Hook Folder for Maintanability
