+++ frontend/src/components/dashboard/DashboardErrorBoundary.jsx
