+++ frontend/src/hooks/transaction/useTransactionData.js
@@ -1,6 +1,6 @@
 import { useQuery, useQueryClient } from '@tanstack/react-query';
-import { transactionAPI, departmentAPI, referrerAPI, revenueAPI, expenseAPI } from '../services/api';
-import { isTestRefunded } from '../utils/transactionUtils';
+import { transactionAPI, departmentAPI, referrerAPI, revenueAPI, expenseAPI } from '../../services/api';
+import { isTestRefunded } from '../../utils/transactionUtils';
 
 /**
  * Custom hook to handle data fetching and basic processing for the Transaction page
