+++ frontend/src/hooks/transaction/useTransactionManagement.js
@@ -1,8 +1,8 @@
 import { useState, useEffect } from 'react';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { transactionAPI } from '../services/api';
+import { transactionAPI } from '../../services/api';
 import { toast } from 'react-toastify';
-import { formatTransactionForDisplay } from '../utils/transactionUtils';
+import { formatTransactionForDisplay } from '../../utils/transactionUtils';
 
 /**
  * Custom hook to manage transaction-related state and actions.
