+++ frontend/src/pages/ActivityLog.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useRef } from 'react'
 import { useLocation } from 'react-router-dom'
 import { Download, RefreshCw, X, Calendar } from 'lucide-react'
 import Sidebar from '../components/Sidebar'
-import useAuth from '../hooks/useAuth'
+import useAuth from '../hooks/auth/useAuth'
 import TabNavigation from '../components/TabNavigation'
 import tabsConfig from '../config/tabsConfig'
 import { activityLogAPI } from '../services/api'
