+++ frontend/src/pages/Login.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginForm from '../components/LoginForm';
+import LoginForm from '../components/auth/LoginForm';
 
 const Login = () => {
   return (
