+++ frontend/src/pages/ReferralManagement.jsx
@@ -2,7 +2,7 @@ import React, { useState, useRef, useEffect } from 'react'
 import { useLocation } from 'react-router-dom'
 import { Download, PlusCircle, X, Calendar, MoreVertical, Edit, Check, AlertTriangle, ArrowUp, ArrowDown } from 'lucide-react'
 import Sidebar from '../components/Sidebar'
-import useAuth from '../hooks/useAuth'
+import useAuth from '../hooks/auth/useAuth'
 import TabNavigation from '../components/TabNavigation'
 import tabsConfig from '../config/tabsConfig'
 import { referrerAPI } from '../services/api'
