+++ frontend/src/pages/Referrals.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useRef, useMemo, useCallback, useEffect } from 'react'
 import Sidebar from '../components/Sidebar'
-import useAuth from '../hooks/useAuth'
+import useAuth from '../hooks/auth/useAuth'
 import { ArrowUp, ArrowDown } from 'lucide-react'
 import { referrerAPI, transactionAPI, departmentAPI } from '../services/api'
 import { useQuery } from '@tanstack/react-query'
