+++ frontend/src/pages/Register.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RegistrationForm from '../components/RegistrationForm';
+import RegistrationForm from '../components/auth/RegistrationForm';
 
 const Register = () => {
   return (
