+++ frontend/src/pages/Settings.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { useNavigate, useLocation } from 'react-router-dom'
 import Sidebar from '../components/Sidebar'
 import TabNavigation from '../components/TabNavigation'
-import useAuth from '../hooks/useAuth'
+import useAuth from '../hooks/auth/useAuth'
 import { Pencil, Key, Users, Save, X } from 'lucide-react'
 import tabsConfig from '../config/tabsConfig'
 import { userAPI } from '../services/api'
