+++ frontend/src/pages/ViewAccounts.jsx
@@ -5,7 +5,7 @@ import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { userAPI } from '../services/api'
 import Sidebar from '../components/Sidebar'
 import TabNavigation from '../components/TabNavigation'
-import useAuth from '../hooks/useAuth'
+import useAuth from '../hooks/auth/useAuth'
 import tabsConfig from '../config/tabsConfig'
 import { ToastContainer, toast } from 'react-toastify'
 import 'react-toastify/dist/ReactToastify.css'
